Exported generateWinCombinations
